refactor: split to single strategy
